requirements.txt
flask==1.0.2
boto3
python-dotenv